ml_service/requirements.txt
#API serving
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.0

#ML / Data
xgboost>=2.0.3
scikit-learn>=1.5.0
numpy>=1.26.4
pandas>=2.2.2
joblib>=1.4.0

#Data collection
riotwatcher==3.3.0         #Riot Games API Python client
pymongo==4.7.0             #Save raw match data to MongoDB
python-dotenv==1.0.1

#Development
pytest==8.2.0
